requirements.txt
streamlit==0.89.0
Pillow==8.4.0
numpy==1.21.5
